pyproject.toml
[project]
name = "iclr-downloader"
version = "0.1.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.5"
dependencies = [
    "openreview-py>=1.47.1",
    "sienna>=0.2.6",
]

[dependency-groups]
dev = [
    "ipython>=8.32.0",
    "mypy>=1.15.0",
    "neovim>=0.3.1",
    "pyright>=1.1.396",
    "pytest>=8.3.4",
    "ruff>=0.9.6",
    "ruff-lsp>=0.0.62",
]
